HTTP POST json to FRED Http-In node Example
From Arduino, post a JSONize data to a server which you create on-the-fly using Node-Red HTTP-In node (in this case i am using FRED).  This is how to do it.  Unfortunately, my Cytron WiFi Shield doesn't perform well on receiving data from Server (happened to MQTT as well).  But as long as sending is working, we are good to publish our data out.